fix: json array stile
