+++ src/parser/grammar.pest
@@ -1,12 +1,12 @@
 WHITESPACE = _{ " " | "\t"}
-URL_SAFE = _{ ASCII_ALPHANUMERIC | "-" | "_" | "." | "~" | "!" | "*" | "(" | ")" | "," | ";" | ":" | "@" | "&" | "=" | "+" | "$" | "/" | "?" | "%" }
+URL_SAFE = _{ ASCII_ALPHANUMERIC | "-" | "_" | "." | "~" | "!" | "*" | "(" | ")" | "," | ";" | ":" | "@" | "&" | "=" | "+" | "$" | "/" | "?" | "%" | "[" | "]" }
 ident = @{ ASCII_ALPHA ~ (ASCII_ALPHANUMERIC | "_")* ~ "'"* }
 int = { ("+" | "-")? ~ ASCII_DIGIT+ }
 num = @{ int ~ ("." ~ ASCII_DIGIT*)? ~ (^"e" ~ int)? }
 hex = { ("0x" | "x") ~ ('0'..'9' | 'a'..'f' | 'A'..'F')+}
 bin = { ("0b" | "b") ~ ("0" | "1")+}
 array = { "[" ~ (ASCII_ALPHANUMERIC | "'" | "\u{0022}" | ",")* ~ "]" }
-quote = { ("'" | "\u{0022}") ~ (URL_SAFE | array)+ ~ (( " " | "\t") ~ (URL_SAFE | array)+)* ~ ("'" | "\u{0022}")}
+quote = { ("'" | "\u{0022}") ~ URL_SAFE+ ~ (( " " | "\t") ~ URL_SAFE+)* ~ ("'" | "\u{0022}")}
 unix = { "unix(" ~ ("'" | "\u{0022}")? ~
 	('0'..'9') ~ ('0'..'9') ~ ('0'..'9')* ~ ('0'..'9')* ~ ("-" | ",") ~
     ('0'..'9') ~ ('0'..'9')* ~ ("-" | ",") ~
